got one button somewhat working
