+++ src/game/editor/prompt.h
@@ -2,7 +2,8 @@
 #define GAME_EDITOR_PROMPT_H
 
 #include <game/client/ui_rect.h>
-#include "game/client/lineinput.h"
+#include <game/client/lineinput.h>
+#include <game/editor/editor_button.h>
 
 #include "component.h"
 
@@ -11,7 +12,7 @@ class CPrompt : public CEditorComponent
 	int m_PromptSelectedIndex = -1;
 
 	std::vector<char *> m_vpCompletePromptList;
-	std::vector<const char *> m_vpFilteredPromptList;
+	std::vector<const CEditorButton*> m_vpFilteredPromptList;
 	CLineInputBuffered<IO_MAX_PATH_LENGTH> m_PromptInput;
 	bool m_Active = false;
